+++ src/test/java/ServiceClasses/ChiSquareTable/ChiSquareTableInterface.java
@@ -1,5 +1,5 @@
 package ServiceClasses.ChiSquareTable;
 
 public interface ChiSquareTableInterface {
-    Long probabilityOf(long ChiTwo, int degreeOfFreedom);
+    Float probabilityOf(Float ChiTwo, int degreeOfFreedom);
 }
